+++ arm-sysregs/examples/log_all.rs
@@ -222,6 +222,8 @@ fn entry(_: u64, _: u64, _: u64, _: u64) -> ! {
     info!("apiakeyhi_el1 = {:?}", arm_sysregs::read_apiakeyhi_el1());
     #[cfg(all(any(test, feature = "fakes", target_arch = "aarch64"), feature = "el1"))]
     info!("apiakeylo_el1 = {:?}", arm_sysregs::read_apiakeylo_el1());
+    #[cfg(all(any(test, feature = "fakes", target_arch = "aarch64"), feature = "el2"))]
+    info!("brbcr_el2 = {:?}", arm_sysregs::read_brbcr_el2());
     #[cfg(any(test, feature = "fakes", target_arch = "arm"))]
     info!("ccsidr = {:?}", arm_sysregs::read_ccsidr());
     #[cfg(any(test, feature = "fakes", target_arch = "arm"))]